requirements.txt
git+https://github.com/mailchimp/mailchimp-marketing-python.git
mollie-api-python